improved error handling logic
