+++ src/error_handlers/__init__.py
@@ -1,0 +1,1 @@
+from error_handlers.error_handlers import _wrap_as_http_exception  # noqa:F401
